// cruscotto/src/app/modules/domande-uniche/dettaglio-istruttoria/models/dettaglioDomandaAccoppiati.ts
export class DatiInterventi {
    input: DettagliCalcolo[];
    output: DettagliCalcolo[];
}

export class DettagliCalcolo {
    codice: any;
    valore: any;
}